Move logic from CcAuthorizationDataBuilder to CheckoutDataBuilder
Remove CcAuthorizationDataBuilder.php
